poshautospam/pyproject.toml
[project]
name = "poshmark-square"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "aiosmtplib>=5.0.0",
    "bs4>=0.0.2",
    "certifi>=2025.11.12",
    "curl-cffi>=0.13.0",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "rich>=14.2.0",
]
